Boat Traversal Changes
Boat Route NavMesh Script:
Boat now follows targets (gameobjects that can be inserted into a list) and adds to index in list to switch to next target after being within 4 meter distance between boat and target.

Boat Tilt script:
made a script that uses the Sine and Cosine functions of Mathf to tilt the boat in the X and Z rotation axis and the y position axis. Intensity of tilt in the various axis can be adjusted in the inspector.