Pointing prompt
Implemented a new script called 'ChoosePromptGesture' (name should be changed), whose function is to add whatever the user is pointing at to a dictionary where the object description of the object is the key, and the corresponding key value the time the user has pointed at the object.
In coherence with this script, changes has also been made in the PointingManager.cs file (script).