'Pointing at' prompt fix
Fixed a problem with the 'rightHandLastSelected' and 'leftHandLastSelected' not being cleared when the dictionary with pointed at points of interest is cleared, resulting in errors in the PointingManager.cs script.

These two variables are used to check for each frame if the user is pointing at an object over a period of time.
Also changed these two variables from gameobjects to strings.