Fading methods
Created a folder with components needed for fading in the scene. The prefab FadeEnd should be placed on the main camera of the XR rig.  When inserted in scene, make sure the "FadeEnd" animator is put on the object if it for some reason is not set on. The script "FadeController" can be used to get the methods that trigger the fade animations. So, if you want the scene to fade when hitting the goal, a script can be made that get the "FadeOut" method from "FadeController", and triggers it when the boat hits the target.