User wave detection + HeadNod, Headshake (WIP)
Wrote a script (WaveTrigger.cs) which together with many other small adjustments and references to scripts enables us to detect when the user is waving.

Tweaked this mentioned implementation together with the user gaze detection so that it registers when the user waves to Ole the fishing villager on the small fishing bridge. This can be used to trigger an animation on Ole so that he waves back to the user, which is easily replicable for other NPCs.

Increased the angle deviation factor for the XR Wave Hand Pose so that it registers more easily.

Also implemented a way for detecting the user headnodding and headshaking, but still needs tweaking. Right now it works with colliders, which are spawned when doing the peace sign with the left hand for debugging purposes.
Should instead be done by looking at the maincamera Y rotation for headshaking, and its X rotation for headnodding.
Zero in these two rotation axes will be looking straight foward like in the camera's scene starting position, which means the minus axis will mean turning the head one direction, and plus axis will mean turning the head the opposite direction.